Fix a regression of windows build issue of undefined symbol
This commit fixes the following build failure.

+ mingw32-make.exe binaries
+ bin/ctr.exe
+ bin/containerd.exe
github.com/containerd/containerd/vendor/github.com/containerd/go-runc
[error]vendor\github.com\containerd\go-runc\runc.go:66:16: undefined: unix.Signal
[error]mingw32-make: *** [Makefile.windows:28: bin/containerd.exe] Error 2
[error]Process completed with exit code 2.

Fixes #65.
Fixes: 421b4ca ("Change the type of PdeathSignal")
